move to test_audit_rest.py
