+++ src/middlewared/middlewared/test/integration/utils/audit.py
@@ -1,4 +1,3 @@
-# -*- coding=utf-8 -*-
 import contextlib
 import pprint
 import time
@@ -9,7 +8,7 @@ from middlewared.test.integration.utils import client
 @contextlib.contextmanager
 def expect_audit_log(entries, *, include_logins=False):
     with client() as c:
-        time.sleep(5)  # FIXME: proper audit log flush
+        time.sleep(5)
 
         existing = c.call("audit.query", {"services": ["MIDDLEWARE"]})
